feat: implement note ownership verification and update note service methods

+++ src/api/notes/handler.js
@@ -15,14 +15,14 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
-      const {
-        title = 'untitled', body, tags,
-      } = request.payload;
+      const {title = 'untitled', body, tags,} = request.payload;
+      const { id:credentialId } = request.auth.credentials
 
       const noteId = await this._service.addNote({
         title,
         body,
         tags,
+        owner: credentialId
       });
 
       const response = h.response({
@@ -56,8 +56,9 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -68,9 +69,10 @@ class NotesHandler {
 
   async getNoteByIdHandler(request, h) {
     try {
-      const {
-        id,
-      } = request.params;
+      const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
       return {
         status: 'success',
@@ -103,9 +105,10 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
-      const {
-        id,
-      } = request.params;
+      const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+      
+      await this._service.verifyNoteOwner(id, credentialId);
 
       await this._service.editNoteById(id, request.payload);
 
@@ -137,9 +140,10 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request, h) {
     try {
-      const {
-        id,
-      } = request.params;
+      const { id } = request.params;
+      const { id: credentialId} = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
 
       return {
